requirements.txt
numpy>=1.26
opencv-python>=4.9
Pillow>=10.0
pytesseract>=0.3
